refactor(menuProduct) : menu 직접참조 제거
