+++ README(2단계).md
@@ -54,6 +54,5 @@
 - [X] 메뉴 상태 변경 메소드 하나의 책임으로 분리
 - [x] menuPrice.compare(menuProducts.getTotalPrice()) getter 대신 메시지 활용
 - [ ] MenuProducts <> menu 양방향 관계 이유
-- [ ] 도메인 서비스의 네이밍은 모델링에 맞춰서 변경해보기
-- 
+- [x] 도메인 서비스의 네이밍은 모델링에 맞춰서 변경해보기
