+++ src/main/java/kitchenpos/menu/tobe/domain/JpaMenuRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, UUID> {
-    @Query("select m from Menu m join m.menuProducts.menuProducts mp where mp.product.id = :productId")
+    @Query("select m from Menu m join m.menuProducts mp where mp.product.id = :productId")
     @Override
     List<Menu> findAllByProductId(@Param("productId") UUID productId);
 }
