+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import jakarta.validation.constraints.NotNull;
-import kitchenpos.product.domain.Product;
+import kitchenpos.products.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.UUID;
