+++ src/test/java/kitchenpos/menu/tobe/domain/MenuPriceTest.java
@@ -3,7 +3,6 @@ package kitchenpos.menu.tobe.domain;
 import kitchenpos.exception.IllegalPriceException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
