+++ src/test/java/kitchenpos/product/tobe/domain/ProductNameTest.java
@@ -3,14 +3,13 @@ package kitchenpos.product.tobe.domain;
 import kitchenpos.exception.IllegalNameException;
 import kitchenpos.infra.FakePurgomalumClient;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.ProductName;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.in;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class ProductNameTest {
